Re-work blob processor integration tests
